\name{benchmark_results}
\alias{benchmark_results}
\docType{data}
\title{Benchmark results}
\description{
Dataset for mnlChoice package. See package vignette for details.
}
\usage{data(benchmark_results)}
\format{Data object. Load with data(benchmark_results) to inspect structure.}
\keyword{datasets}

