\name{brier_score}
\alias{brier_score}
\title{Brier score}
\description{
Function from mnlChoice package. See source code in R/brier_score.R for detailed documentation.
}
\usage{
brier_score(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
brier_score()
}
}

