\name{check_mnp_convergence}
\alias{check_mnp_convergence}
\title{Check mnp convergence}
\description{
Function from mnlChoice package. See source code in R/check_mnp_convergence.R for detailed documentation.
}
\usage{
check_mnp_convergence(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
check_mnp_convergence()
}
}

