\name{compare_mnl_mnp}
\alias{compare_mnl_mnp}
\title{Compare mnl mnp}
\description{
Function from mnlChoice package. See source code in R/compare_mnl_mnp.R for detailed documentation.
}
\usage{
compare_mnl_mnp(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
compare_mnl_mnp()
}
}

