\name{convergence_report}
\alias{convergence_report}
\title{Convergence report}
\description{
Function from mnlChoice package. See source code in R/convergence_report.R for detailed documentation.
}
\usage{
convergence_report(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
convergence_report()
}
}

