\name{decision_framework}
\alias{decision_framework}
\title{Decision framework}
\description{
Function from mnlChoice package. See source code in R/decision_framework.R for detailed documentation.
}
\usage{
decision_framework(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
decision_framework()
}
}

