\name{evaluate_by_estimand}
\alias{evaluate_by_estimand}
\title{Evaluate by estimand}
\description{
Function from mnlChoice package. See source code in R/evaluate_by_estimand.R for detailed documentation.
}
\usage{
evaluate_by_estimand(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
evaluate_by_estimand()
}
}

