\name{evaluate_performance}
\alias{evaluate_performance}
\title{Evaluate performance}
\description{
Function from mnlChoice package. See source code in R/evaluate_performance.R for detailed documentation.
}
\usage{
evaluate_performance(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
evaluate_performance()
}
}

