\name{fit_mnp_safe}
\alias{fit_mnp_safe}
\title{Fit mnp safe}
\description{
Function from mnlChoice package. See source code in R/fit_mnp_safe.R for detailed documentation.
}
\usage{
fit_mnp_safe(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
fit_mnp_safe()
}
}

