\name{flexible_mnl}
\alias{flexible_mnl}
\title{Flexible mnl}
\description{
Function from mnlChoice package. See source code in R/flexible_mnl.R for detailed documentation.
}
\usage{
flexible_mnl(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
flexible_mnl()
}
}

