\name{functional_form_test}
\alias{functional_form_test}
\title{Functional form test}
\description{
Function from mnlChoice package. See source code in R/functional_form_test.R for detailed documentation.
}
\usage{
functional_form_test(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
functional_form_test()
}
}

