\name{generate_choice_data}
\alias{generate_choice_data}
\title{Generate choice data}
\description{
Function from mnlChoice package. See source code in R/generate_choice_data.R for detailed documentation.
}
\usage{
generate_choice_data(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
generate_choice_data()
}
}

