\name{interpret_convergence_failure}
\alias{interpret_convergence_failure}
\title{Interpret convergence failure}
\description{
Function from mnlChoice package. See source code in R/interpret_convergence_failure.R for detailed documentation.
}
\usage{
interpret_convergence_failure(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
interpret_convergence_failure()
}
}

