\name{mnl_mnp_benchmark}
\alias{mnl_mnp_benchmark}
\docType{data}
\title{Mnl mnp benchmark}
\description{
Dataset for mnlChoice package. See package vignette for details.
}
\usage{data(mnl_mnp_benchmark)}
\format{Data object. Load with data(mnl_mnp_benchmark) to inspect structure.}
\keyword{datasets}

