\name{model_summary_comparison}
\alias{model_summary_comparison}
\title{Model summary comparison}
\description{
Function from mnlChoice package. See source code in R/model_summary_comparison.R for detailed documentation.
}
\usage{
model_summary_comparison(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
model_summary_comparison()
}
}

