\name{plot_convergence_rates}
\alias{plot_convergence_rates}
\title{Plot convergence rates}
\description{
Function from mnlChoice package. See source code in R/plot_convergence_rates.R for detailed documentation.
}
\usage{
plot_convergence_rates(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
plot_convergence_rates()
}
}

