\name{plot_recommendation_regions}
\alias{plot_recommendation_regions}
\title{Plot recommendation regions}
\description{
Function from mnlChoice package. See source code in R/plot_recommendation_regions.R for detailed documentation.
}
\usage{
plot_recommendation_regions(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
plot_recommendation_regions()
}
}

