\name{power_analysis_mnl}
\alias{power_analysis_mnl}
\title{Power analysis mnl}
\description{
Function from mnlChoice package. See source code in R/power_analysis_mnl.R for detailed documentation.
}
\usage{
power_analysis_mnl(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
power_analysis_mnl()
}
}

