\name{publication_table}
\alias{publication_table}
\title{Publication table}
\description{
Function from mnlChoice package. See source code in R/publication_table.R for detailed documentation.
}
\usage{
publication_table(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
publication_table()
}
}

