\name{quantify_model_choice_consequences}
\alias{quantify_model_choice_consequences}
\title{Quantify model choice consequences}
\description{
Function from mnlChoice package. See source code in R/quantify_model_choice_consequences.R for detailed documentation.
}
\usage{
quantify_model_choice_consequences(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
quantify_model_choice_consequences()
}
}

