\name{quick_decision}
\alias{quick_decision}
\title{Quick decision}
\description{
Function from mnlChoice package. See source code in R/quick_decision.R for detailed documentation.
}
\usage{
quick_decision(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
quick_decision()
}
}

