\name{recommend_model}
\alias{recommend_model}
\title{Recommend model}
\description{
Function from mnlChoice package. See source code in R/recommend_model.R for detailed documentation.
}
\usage{
recommend_model(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
recommend_model()
}
}

