\name{required_sample_size}
\alias{required_sample_size}
\title{Required sample size}
\description{
Function from mnlChoice package. See source code in R/required_sample_size.R for detailed documentation.
}
\usage{
required_sample_size(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
required_sample_size()
}
}

