\name{run_benchmark_simulation}
\alias{run_benchmark_simulation}
\title{Run benchmark simulation}
\description{
Function from mnlChoice package. See source code in R/run_benchmark_simulation.R for detailed documentation.
}
\usage{
run_benchmark_simulation(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
run_benchmark_simulation()
}
}

