\name{sample_size_table}
\alias{sample_size_table}
\title{Sample size table}
\description{
Function from mnlChoice package. See source code in R/sample_size_table.R for detailed documentation.
}
\usage{
sample_size_table(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
sample_size_table()
}
}

