\name{simulate_dropout_scenario}
\alias{simulate_dropout_scenario}
\title{Simulate dropout scenario}
\description{
Function from mnlChoice package. See source code in R/simulate_dropout_scenario.R for detailed documentation.
}
\usage{
simulate_dropout_scenario(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
simulate_dropout_scenario()
}
}

