\name{substitution_matrix}
\alias{substitution_matrix}
\title{Substitution matrix}
\description{
Function from mnlChoice package. See source code in R/substitution_matrix.R for detailed documentation.
}
\usage{
substitution_matrix(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
substitution_matrix()
}
}

