\name{test_iia}
\alias{test_iia}
\title{Test iia}
\description{
Function from mnlChoice package. See source code in R/test_iia.R for detailed documentation.
}
\usage{
test_iia(...)
}
\arguments{
\item{...}{Function arguments. See source file for parameter descriptions.}
}
\value{
Function return value. See source file for details.
}
\details{
This function is part of the mnlChoice package for comparing MNL and MNP models.
Full documentation is available in the R source file and package vignette.
}
\examples{
# See package vignette for examples
\dontrun{
test_iia()
}
}

