\name{validation_results}
\alias{validation_results}
\docType{data}
\title{Validation results}
\description{
Dataset for mnlChoice package. See package vignette for details.
}
\usage{data(validation_results)}
\format{Data object. Load with data(validation_results) to inspect structure.}
\keyword{datasets}

